HSEARCH-5073 Add minimum_should_match to simple query string predicate
