+++ engine/src/main/java/org/hibernate/search/engine/search/predicate/spi/QueryStringPredicateBuilder.java
@@ -8,7 +8,7 @@ package org.hibernate.search.engine.search.predicate.spi;
 
 import org.hibernate.search.engine.search.common.RewriteMethod;
 
-public interface QueryStringPredicateBuilder extends MinimumShouldMatchBuilder, CommonQueryStringPredicateBuilder {
+public interface QueryStringPredicateBuilder extends CommonQueryStringPredicateBuilder {
 
 	void allowLeadingWildcard(boolean allowLeadingWildcard);
