+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/search/predicate/QueryStringPredicateSpecificsIT.java
@@ -179,96 +179,6 @@ class QueryStringPredicateSpecificsIT extends AbstractBaseQueryStringPredicateSp
 		);
 	}
 
-	@ParameterizedTest
-	@MethodSource("minimumShouldMatch")
-	void minimumShouldMatch(String phrase, int min, String docId, String[] otherDocIds) {
-		StubMappingScope scope = index.createScope();
-		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
-
-		SearchQuery<DocumentReference> query = scope.query()
-				.where( f -> f.queryString()
-						.field( absoluteFieldPath )
-						.matching( phrase )
-						.minimumShouldMatchNumber( min ) )
-				.toQuery();
-		if ( docId == null ) {
-			assertThatQuery( query )
-					.hasNoHits();
-		}
-		else {
-			assertThatQuery( query )
-					.hasDocRefHitsAnyOrder( index.typeName(), docId, otherDocIds );
-		}
-	}
-
-	@ParameterizedTest
-	@MethodSource("minimumShouldMatch")
-	void minimumShouldMatch_dsl(String phrase, int min, String docId, String[] otherDocIds) {
-		StubMappingScope scope = index.createScope();
-		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
-
-		SearchQuery<DocumentReference> query = scope.query()
-				.where( f -> f.queryString()
-						.field( absoluteFieldPath )
-						.matching( phrase )
-						.minimumShouldMatch().ifMoreThan( 0 )
-						.thenRequireNumber( min )
-						.end() )
-				.toQuery();
-		if ( docId == null ) {
-			assertThatQuery( query )
-					.hasNoHits();
-		}
-		else {
-			assertThatQuery( query )
-					.hasDocRefHitsAnyOrder( index.typeName(), docId, otherDocIds );
-		}
-	}
-
-	public static List<? extends Arguments> minimumShouldMatch() {
-		// "Here I was, feeding my panda, and the crowd had no word."
-		return List.of(
-				Arguments.of( "feeding my panda", 1, DOCUMENT_1, new String[] { DOCUMENT_3 } ),
-				Arguments.of( "feeding my panda", 2, DOCUMENT_1, new String[] { } ),
-				Arguments.of( "feeding my panda", 3, DOCUMENT_1, new String[] { } ),
-				Arguments.of( "feeding my panda crowd had", 5, DOCUMENT_1, new String[] { } ),
-				Arguments.of( "some other panda less matches", 5, null, new String[] { } )
-		);
-	}
-
-	@ParameterizedTest
-	@MethodSource("minimumShouldMatchPercent")
-	void minimumShouldMatch_percent(String phrase, int min, String docId, String[] otherDocIds) {
-		StubMappingScope scope = index.createScope();
-		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
-
-		SearchQuery<DocumentReference> query = scope.query()
-				.where( f -> f.queryString()
-						.field( absoluteFieldPath )
-						.matching( phrase )
-						.minimumShouldMatchPercent( min ) )
-				.toQuery();
-		if ( docId == null ) {
-			assertThatQuery( query )
-					.hasNoHits();
-		}
-		else {
-			assertThatQuery( query )
-					.hasDocRefHitsAnyOrder( index.typeName(), docId, otherDocIds );
-		}
-	}
-
-	public static List<? extends Arguments> minimumShouldMatchPercent() {
-		// "Here I was, feeding my panda, and the crowd had no word."
-		return List.of(
-				Arguments.of( "feeding my panda", 34, DOCUMENT_1, new String[] { DOCUMENT_3 } ),
-				Arguments.of( "feeding my panda", 67, DOCUMENT_1, new String[] { } ),
-				Arguments.of( "feeding my panda", 100, DOCUMENT_1, new String[] { } ),
-				Arguments.of( "feeding my panda crowd had", 100, DOCUMENT_1, new String[] { } ),
-				Arguments.of( "some other panda less matches", 100, null, new String[] { } )
-		);
-	}
-
 	@Test
 	void enablePositionIncrements_disabled() {
 		StubMappingScope scope = index.createScope();
